in.zoom {zoom}	R Documentation

Direct access to zoom functionalities.

Description

Direct selection of a zoom method of the "session" type. Possibly of use in scripts?

allow interactive in/out zoom in "session" mode

Usage

  in.zoom(...)

  move.to.click.zoom(...)

  inout.zoom(...)

  out.zoom(...)

  set.zoom(...)

  sq.zoom(...)

Arguments

...
Extra arguments to zoomplot.zoom.

Value

NULL

Note

Each function starts a different interactive sequence

inout.zoom(): left click within bounds zooms in, outside bounds zoom out

move.to.click.zoom(): center plot arround left click

in.zoom(): each left click zooms in

out.zoom(): each left click zooms out

set.zoom(): ask for a magnification factor

sq.zoom(): allow to click on the two corners of the desired region to zoom on

Author(s)

Corentin M. Barbu

See Also

zm(), session.zoom().

[Package zoom version 2.0.4 Index]