session.zoom {zoom}	R Documentation

Opening of an interactive zoom/navigate session.

Description

To launch an interactive session you should use zm() but if you are sure of your device you can launch directly one of these functions.

Usage

  session.zoom(...)

  navigation.zoom(...)

Arguments

...
Everything that can be accepted by sq.zoom.

Details

session.zoom launch an interactive console menu to navigate a plot.

navigation.zoom allows to interactively navigate a plot with the mouse.

Value

Returns the final plot, as saved by recordPlot().

Author(s)

Corentin M. Barbu, Sebastian Gibb

See Also

zm().

Examples

## Not run: 
plot(rnorm(100),rnorm(100))
session.zoom()

## End(Not run)

[Package zoom version 2.0.4 Index]