zoom-package {zoom}	R Documentation

The zoom Package

Description

A spatial data visualization tool.

Details

Package: 	 zoom

Type: 	 Package

Version: 	 2.0.4

Date: 	 2013-10-14

Depends: 	 R (>= 2.10.0)

Encoding: 	 UTF-8

License: 	 GPL (>= 3)

LazyLoad: 	 yes

URL: 	 https://github.com/cbarbu/R-package-zoom

zm(), called with any active plot allow to enter an interactive session to zoom/navigate any plot. The development version, as well as binary releases can be found at https://github.com/cbarbu/R-package-zoom

Author(s)

Corentin M Barbu corentin.barbu@gmail.com, with contributions from Sebastian Gibb mail@sebastiangibb.de

[Package zoom version 2.0.4 Index]