zoomplot.zoom {zoom}	R Documentation

Central low level function of the zoom package.

Description

This function allow to replot the current or a saved plot with specific boundaries, magnification factor and possibly arround a user defined x/y.

Usage

  zoomplot.zoom(xlim = NULL, ylim = NULL, fact = NULL,
    rp = NULL, x = NULL, y = NULL, xlimfn = NULL,
    ylimfn = NULL, ...)

Arguments

xlim
A vector with min and max x

ylim
A vector with min and max y

fact
A scalar giving the magnification factor (>1 brings you closer)

rp
A previously recorded plot with recordPlot(). With all the corresponding warnings in ?recordPlot.

x
x of a fix point when rescaling, by default the center.

y
y of a fix point when rescaling, by default the center.

xlimfn
a function using x, y and/or fact to generate new x lim if NULL and xlim/ylim not given will use multipancPoint

ylimfn
a function using x, y and/or fact to generate new y lim, if NULL will use xlimfn

...
Additional parameters not implemented, just in case.

Details

This function is not necessarily easy to use by hand. It is designed to work well when called from higher level functions. End user should always use zm().

Value

Not guaranted for now.

Note

This function is the heart of the zoom package and the one that can be. affected by R version changes. It is inspired by the zoomplot function in TeachingDemos package

Author(s)

Corentin M. Barbu

See Also

zm, in.zoom

Examples

plot(rnorm(1000),rnorm(1000))
zoomplot.zoom(fact=2,x=0,y=0)

[Package zoom version 2.0.4 Index]